test_requirements.txt
keras
h5py
mxnet
Pillow
requests
torch
tensorflow